requirements.txt
langchain>=1.0.0
langchain-openai>=0.3.0
langgraph>=1.0.0
pydantic>=2.0.0
python-dotenv